requirements.txt
pytesseract
Pillow
transformers
torch
exifread
